Separate web registerWith plugin
